Make legislation view reusable in liiweb
