--- lawlibrary/templates/lawlibrary/legislation_list.html
@@ -1,37 +1,0 @@
-{% extends "peachjam/layouts/main.html" %}
-{% load i18n %}
-{% block title %} {% trans 'Legislation' %} {% endblock %}
-
-{% block page-content %}
-  <section class="pb-5">
-    <div class="container">
-      {% block breadcrumbs %}
-      {% endblock %}
-
-      <h1 class="my-4">{% trans 'National Legislation' %}</h1>
-
-      {% include 'lawlibrary/_legislation_alert.html' %}
-
-      <nav class="nav nav-tabs mb-3 border-bottom">
-        <li class="nav-item">
-          <a class="nav-link {% if view.variant == 'current' %}active{% endif %}" href="{% url 'legislation_list' %}">Current legislation</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link {% if view.variant == 'repealed' %}active{% endif %}" href="{% url 'legislation_list_repealed' %}">Repealed legislation</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link {% if view.variant == 'all' %}active{% endif %}" href="{% url 'legislation_list_all' %}">All legislation</a>
-        </li>
-      </nav>
-
-      {% if view.variant == 'repealed' %}
-        <div class="alert alert-danger">
-          You are viewing repealed legislation which is no longer in force.
-        </div>
-      {% endif %}
-
-      <div data-vue-component="LegislationTable" data-show-side-facets></div>
-      {{ legislation_table|json_script:"legislation-table" }}
-    </div>
-  </section>
-{% endblock %}
