+++ lawlibrary/templates/liiweb/legislation_list.html
@@ -1,0 +1,11 @@
+{% extends "liiweb/legislation_list.html" %}
+{% load i18n %}
+
+{% block page-heading %}
+  <h1 class="my-4">{% trans 'National Legislation' %}</h1>
+{% endblock %}
+
+{% block legislation-nav %}
+  {% include 'lawlibrary/_legislation_alert.html' %}
+  {{ block.super }}
+{% endblock %}
