+++ liiweb/views/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
 from .home import *
+from .legislation import *
